requirement.txt
asgiref==3.2.3
autopep8==1.4.4
dataclasses==0.6
Django==2.1.2
django-environ==0.4.5
django-rest-framework-mongoengine==3.4.0
djangorestframework==3.11.0
dnspython==1.16.0
mongodbforms==0.3
mongoengine==0.18.2
psycopg2-binary==2.8.4
pycodestyle==2.5.0
pymongo==3.10.0
pytz==2019.3
six==1.13.0
sqlparse==0.2.4
